Hopefully final .grpc. to .proto. rename.
